requirements.txt
# Core LangChain
langchain==1.0.4
langchain-huggingface==0.0.9
langchain-chroma==0.1.0
langchain-community==0.0.30
langchain-google-genai==0.0.11

# Vector store
chromadb==0.3.26

# Transformers & embeddings
torch==2.2.0
sentence-transformers==2.2.2
transformers==4.35.0

# PDF / OCR
pymupdf==1.22.5
pdf2image==1.16.3
pytesseract==0.3.10
pillow==10.1.0

# Environment variables
python-dotenv==1.0.1

# Streamlit UI
streamlit==1.26.0
